Release designate-dashboard for stable/train
This release picks up new commits to designate-dashboard since
the last release from stable/train.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a  stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

Change-Id: I371efe15b92f3d9d38abb88649798e7d4931dd09
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>