Release monasca-agent for stable/train
This release picks up new commits to monasca-agent since
the last release from stable/train.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a  stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

Change-Id: I4b018444e12f8ffa6e9489873e384288e413673c
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>