Cleanup old independent releases
A lot of OpenStack deliverables start their existence as cycle-independent,
as it is what our docs still recommend for initial releases. Most of them
quickly turn into cycle-following deliverables, however their early
lives as independent deliverables produce a confusing trail at [1].

This page seems to imply (for example) that Cyborg is a
cycle-independent deliverable and that its most recent version is 0.2.0.
This can be very confusing, and displaying long-forgotten early releases
on releases.o.o is not worth it.

Note: cycle-following deliverables that become independent actually do
not generate the same confusion. They stop being listed on the
coordinated release pages and their listing on [1] is accurate. It's the
independent deliverables that become tied to the cycle that result in
confusing information.

This change removes deliverable files for deliverables that had a couple
of independent releases at the start of their lives but are now
following the release cycle. Those releases can still be found on
tarballs.o.o and their tags still exist in their history. We just stop
displaying them on releases.o.o.

This allows the "Series-Independent Projects" page[1] to only display
deliverables that actually *are* released in a cycle-independent manner.

[1] https://releases.openstack.org/independent.html

Change-Id: Ic99b469cb15e87ea9b6e3505d2280d45cd224348