Release python-novaclient for victoria-1 milestone
This is a library release for python-novaclient for the victoria-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

Change-Id: Ia54a9b1551a77b5eaf1eb925e7c465c99d0c30d6
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>