Move the "Adding a deliverable" section
Information about adding a deliverable should live in the
general usage documentation, not the release models page.

Also stop advising projects to release independently if they
get added after the second milestone of a series, as it creates
issues down the road with inconsistent stale pages in the
'Independent' section.

Change-Id: I9319656b62b063d637737ad8667d3099f179fb58