Octavia: EOL Rocky
EOL Rocky has been discussed on the openstack-discuss list [1] and on
weekly Octavia project meetings.

Testing on this branch has become less and less stable (Python 2 drop in
dependent projects, broken devstack, etc). Community's interest in this
branch is very slim to non-existent, with no volunteers to keep it
afloat.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2020-April/014057.html

Change-Id: Ic03af87408a0e8caefdd9b1d0b8f50d31c0991ae