Release ironic-inspector for stable/stein
This release picks up new commits to ironic-inspector since
the last release from stable/stein.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a  stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

This is not a release team activity, it is just to help out with
stable releases. If there is no response to the patch, we will
assume a release is not needed and abandon the patch after one
week.

Change-Id: I1d092f4eb08055a9ad511b23795956b16c1f5cbd
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>