Release kuryr-libnetwork for stable/stein
This release picks up new commits to kuryr-libnetwork since
the last release from stable/stein.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a  stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

This is not a release team activity, it is just to help out with
stable releases. If there is no response to the patch, we will
assume a release is not needed and abandon the patch after one
week.

Change-Id: I3999d0f407e5cb27e7da1c2edbc6ed79f379625d
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>