Fix list_unreleased_changes when repo doesn't exist or haven't been released
Indeed during my previous implementation of this script, I missed to handle
the cases where repo is not found or has not yet been released when users
ask for a standard output. For yaml and json is not an issue as we dump
raw dict at the asked format (json or yaml).

These changes fix that.

Also to stay closer of the previous implementation of this tools (the
shell version), these changes add:
- colored outputs
- compared revisions (even when no diff are found,
  like in the previous shell version).

Change-Id: I74f0f8c7071fbf2c8272fc002f28b793ff8e672d