Victoria-2 Release for python-karborclient
This is the victoria milestone-2 release for python-karborclient.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by August 6.

Change-Id: If643d49976e8270d8c8d1653e1a3055919eff21f
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>