Switch magnum-ui to cycle-with-rc
Deliverables following the cycle-with-intermediary model are expected
to release multiple times throughout the development cycle. Since
magnum-ui only released once last cycle, and has not done a
release yet this cycle, this proposes changing the release model to
cycle-with-rc so the expectation is only that it will release an RC by
the end of the cycle prior to doing the final release.

This is just a suggestion for the team to consider. The important point
is past the milestone-2 point, downstream planning starts happening
for what will be included in the release. We just want to try to avoid
the situation where downstream expects something to be included, but we
need to unexpectedly drop it from the coordinated release late in the
cycle.

For the team, please respond to this patch in one of the following ways:

* -1 this patch and follow up with an intermediary release
* +1 this patch and we will switch over the release model
* -1 this patch, but acknowledge a release will be done prior to the
  RC1 deadline.

Change-Id: I36ad5513dc8dc5fa5dcc3003d7b32b46e88cb6ac
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>