release oslo.messaging 12.1.4 - 2020-09-01
I released a bugfix version even if a new option have been added [1]
because 12.2.x is already consumed by victoria.

Although this adds a new option, it's just a gate on a
bug fix to ensure we don't have unintended side effects. That actually
makes this backport safer than if it just unconditionally enabled the
fix.

[1] https://review.opendev.org/#/c/747366/

Change-Id: Ifed5068d5762d241233e84bf419d586f9a2b39ed