Release final metalsmith for Victoria
This proposed a release for metalsmith to get any final merged changes
out prior to the lib freeze for Victoria going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, Sept 7.

Change-Id: Iefb6a401b9df100d301a0fb7bb94641b8b07fe19
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>