Release final sushy for Victoria
This proposed a release for sushy to get any final merged changes
out prior to the lib freeze for Victoria going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, Sept 7.

Change-Id: I81144f983b64fd0ad2ac215b08fc9ce38a023559
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>