Release final python-brick-cinderclient-ext for victoria
Final victoria client lib release for the September 10 deadline.

Please leave +1 if we are all set and can proceed. Leave a -1 if we
need to wait a little bit longer for something final to loand,
then update with a better commit hash to use for the release when
ready.

Patches with no response will be approved by September 14.

Change-Id: Id06b968988677b4a12fcf5bb8a8becdcea0d4dc9
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>