Release final python-cyborgclient for victoria
Final victoria client lib release for the September 10 deadline.

Please leave +1 if we are all set and can proceed. Leave a -1 if we
need to wait a little bit longer for something final to loand,
then update with a better commit hash to use for the release when
ready.

Patches with no response will be approved by September 14.

Change-Id: I6cc131562942535b79f8e8a1517a7b5b2b1b0b69
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>