Release final python-manilaclient for victoria
Final victoria client lib release for the September 10 deadline.

Please leave +1 if we are all set and can proceed. Leave a -1 if we
need to wait a little bit longer for something final to loand,
then update with a better commit hash to use for the release when
ready.

Patches with no response will be approved by September 14.

Change-Id: Iacc599380a30b4c7c004af78fa8c9a184b7f1d73
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>