Release final python-masakariclient for victoria
Final victoria client lib release for the September 10 deadline.

Please leave +1 if we are all set and can proceed. Leave a -1 if we
need to wait a little bit longer for something final to loand,
then update with a better commit hash to use for the release when
ready.

Patches with no response will be approved by September 14.

Change-Id: If63b499632ad8b28732458b26f680a839a225646
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>