Release final python-muranoclient for victoria
Final victoria client lib release for the September 10 deadline.

Please leave +1 if we are all set and can proceed. Leave a -1 if we
need to wait a little bit longer for something final to loand,
then update with a better commit hash to use for the release when
ready.

Patches with no response will be approved by September 14.

Change-Id: I45685b7aa259d63480ce0c5273bd4338655271b8
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>