Release final python-octaviaclient for victoria
Final victoria client lib release for the September 10 deadline.

Please leave +1 if we are all set and can proceed. Leave a -1 if we
need to wait a little bit longer for something final to loand,
then update with a better commit hash to use for the release when
ready.

Patches with no response will be approved by September 14.

Change-Id: I2dc6bec2699427ca630958252bedc978223157c2
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>