Release final python-searchlightclient for victoria
Final victoria client lib release for the September 10 deadline.

Please leave +1 if we are all set and can proceed. Leave a -1 if we
need to wait a little bit longer for something final to loand,
then update with a better commit hash to use for the release when
ready.

Patches with no response will be approved by September 14.

Change-Id: I3d258fa6b473d84f91cb6126018070c2e66d4b06
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>