Release final python-swiftclient for victoria
Final victoria client lib release for the September 10 deadline.

Please leave +1 if we are all set and can proceed. Leave a -1 if we
need to wait a little bit longer for something final to loand,
then update with a better commit hash to use for the release when
ready.

Patches with no response will be approved by September 14.

Change-Id: I889c55dca2d9be5adf7cee3c4b7058575ef2402b
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>