Release final python-troveclient for victoria
Final victoria client lib release for the September 10 deadline.

Please leave +1 if we are all set and can proceed. Leave a -1 if we
need to wait a little bit longer for something final to loand,
then update with a better commit hash to use for the release when
ready.

Patches with no response will be approved by September 14.

Change-Id: I6a1f8768c20847b18e7b9021d9711357bfb942de
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>