Proposing victoria RC1 for barbican
As a courtesy for project teams, on RC1 week the release management
team proposes a RC1 release for all cycle-with-rc deliverables,
based on HEAD. That RC1 will double as the branch point for the
stable/victoria branch for barbican.

Release liaison or PTL should +1 this proposal if they are happy
with it, amend this proposal (for example with a more recent SHA)
to match their needs, or -1 to block it until release-critical
issues are all fixed.

Change-Id: I9033c82d1a14db7e40ad3f3ad117ac264c84efdd
Signed-off-by: Thierry Carrez <thierry@openstack.org>