Proposing victoria RC1 for networking-bgpvpn
As a courtesy for project teams, on RC1 week the release management
team proposes a RC1 release for all cycle-with-rc deliverables,
based on HEAD. That RC1 will double as the branch point for the
stable/victoria branch for networking-bgpvpn.

Release liaison or PTL should +1 this proposal if they are happy
with it, amend this proposal (for example with a more recent SHA)
to match their needs, or -1 to block it until release-critical
issues are all fixed.

Change-Id: Id5e99fc14b2bd2e6c6daa70a1452c9b9eb0517d4
Signed-off-by: Thierry Carrez <thierry@openstack.org>