Proposing victoria RC1 for sahara-plugin-ambari
As a courtesy for project teams, on RC1 week the release management
team proposes a RC1 release for all cycle-with-rc deliverables,
based on HEAD. That RC1 will double as the branch point for the
stable/victoria branch for sahara-plugin-ambari.

Release liaison or PTL should +1 this proposal if they are happy
with it, amend this proposal (for example with a more recent SHA)
to match their needs, or -1 to block it until release-critical
issues are all fixed.

Change-Id: Iccfd1322e1cdbdd2ca31fc45c5a9affa3292bdc8
Signed-off-by: Thierry Carrez <thierry@openstack.org>