Victoria final releases for cycle-with-rc projects
This tags the final release for all cycle-with-rc projects for Victoria.
While not required, having PTLs ack this patch would be nice to include
that record in the merged metadata.

Release activity should now be frozen until this is processed on Oct 14.
Any bugfixes merged to stable/victoria can be queued up and ready to go
as a follow on stable release after the coordinated release date.

There are different diff-start values for each project. This is used
when generating release notes to denote the point where the
stable/ussuri branch started to diverge from what is included on
stable/victoria. For some, this will be the final Ussuri major version,
but if multiple RCs were done, this may show the initial RC1 release
version. This is normal.

Change-Id: I71d78cf386b960c20f6f453fb13c263d4c9b2e18
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>