Adding Daniel Bengtsson as release liaison for oslo
Oslo team will soon adopt the DPL governance model [1], according to
this governance switch we need to define required roles and the
release liaison is one of these roles [2].

During the last oslo meeting we reached a consensus and we decided to
create groups of paired liaison [3], especially for release and TaCT SIG [4].

Daniel has been volunteer and proposed himself on this role.
He proposed to help us on the release topic.

Daniel will be my partner on this role during the Wallaby cycle and he will
be mentored by myself to transfer the needed knowledges to him [4].

[1] https://governance.openstack.org/tc/resolutions/20200803-distributed-project-leadership.html
[2] https://governance.openstack.org/tc/resolutions/20200803-distributed-project-leadership.html#required-roles
[3] http://eavesdrop.openstack.org/meetings/oslo/2020/oslo.2020-10-12-15.00.log.txt
[4] https://wiki.openstack.org/wiki/Oslo#Project_Leadership_Liaisons

Change-Id: Idcf2251504da22696402e81cbf71475084201637