init-series: Skip unrelease comment for trailing
Cycle trailing deliverables are often not released by the time we
initialize the next series. We add a comment to deliverable files during
their creation to note if they were not released in the last cycle to
help highlight orphaned deliverables. This can be misleading for cycle
trailing deliverables, since usually they do end up being released, but
after we've created the file and included the comment.

To make this comment more useful, this removes its inclusion for
trailing projects.

Change-Id: I010dadc2ca9026248e7f9dfc89d257afdf97be72
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>