Automate review inbox dashboard link
We have gerrit-dash-creator and https://tiny.cc/ReleaseInbox, but both
require manual steps to be taken by someone each time the status of a
series changes. Since we already capture series status data in a format
that is easily parsable, let's just add a sphinx directive to make this
automatic.

Change-Id: I05b2a65b52809d249bfb299164d624f6de5bff02
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>