List the Wallaby cycle signing key as current
This sets the end date for the Victoria key, adds the Wallaby key
starting today and links to a minimal export of the corresponding
public key and its details on the keyserver network. This can be
amended with a new patchset or a follow-up change to correct the
included dates after the key material change (Depends-On below)
merges and goes into effect.

Change-Id: I97ca5ee0baf20a9470189fa621a1fa30115aeb15
Depends-On: https://review.opendev.org/760364