Moving oslo projects to independent model
As discussed on the ML many oslo projects are now stable enough to be
adopt the independent release model.

It will help us o reduce the amount of efforts needed to backport fixes to
stable branches.

oslo.upgradecheck, tooz, stevedore, and osprofiler are finally ignored as they
needs some releases for Wallaby. They was originally planed to be
transitioned too.

http://lists.openstack.org/pipermail/openstack-discuss/2020-November/018905.html

Change-Id: I6ac8b52349d631171647c89c8ef18cd6a9b16808