[Tempest Plugins] Tag stein-last
Stein branch is in Extended Maintenance now[1]. All Tempest
plugins are branchless which means master version of Tempest
and its plugins is used to test the supported stable branches.

Once stable branch is moved to EM state then, Tempest and its
plugins compatible tag needs to be released so that we can
keep testing the EM stable branches with this tag once master
Tempest and its plugins are not compatible[2].

This tag will help for testing the stable/stein with single
compatible tag of Tempest and its plugins.

Tagging 'stein-last' with the commit used for 0.6.0.

More detail on compatible version policy for Tempest and its
plugins is here
- https://docs.openstack.org/tempest/latest/tempest_and_plugins_compatible_version_policy.html

[1] https://releases.openstack.org/stein/index.html
[2] https://docs.openstack.org/tempest/latest/stable_branch_support_policy.html

Change-Id: I95ef209e4f1a6fabd01eab2d796c5b933f1c3b62