Allow to tag projects as "last" on EM branches
For testing for Extended Maintenance branches with branchless
Tempest and Tempest plugins, we need to release the new tag
for Tempest and its plugins as last compatible version for that
Extended Maintenance branches.

This commits adds the documentation and validation machinery changes about
that tag $series-last which will be used for branchless testing tooling.

Change-Id: I799e8e637a54a46fd7ca74dd568ea2c7506fa32d
Depends-On: https://review.opendev.org/c/openstack/project-team-guide/+/769821