Wallaby-2 Release for python-novaclient
This is the Wallaby milestone-2 release for python-novaclient.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If you are happy with the current release hash, please +1 this patch
and we will process a release. If you need to get some important
changes merged first, leave a -1 and update the patch with a new
commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed by January 21.

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I8ad78d906fc03a5bb46a75c120ae21950b64403a