Release Octavia 5.1.0 for Train
This release includes changes to the requirements and lower-constraints
file because of the new pip release. It also includes a bug fix where a
new octavia.conf configuration option (default_connection_limit) with a
sane default was added to fix an issue where HAProxy 1.8 consumes much
more memory and crashes the amphora load balancer. Because of these two
points, we are bumping the minor version.

Change-Id: Ib7a951064edb7b8a89bbf9c9ad69f6cd19e2f543