Ensure that we can build sdist from requested tags
Adding validation check to ensure that we are able to build sdist from
the latest tag.

This will allow us to detect sooner when we will try to add a tag to a branch
which is older than another tag already on that branch/in that branch's history

Indeed we recently experienced similar issue [1] with os-collect-config and
tripleo-ipsec where projects hadn't been branched for ussuri [2] and where
we tried to create a new ussuri release [3] on the same hash that an
existing wallaby tag.

Projects have been branched [3] after tags creation [2].

[1] http://lists.openstack.org/pipermail/openstack-discuss/2021-January/020112.html
[2] https://opendev.org/openstack/releases/commit/688914aa29745bd4f445670581f79760e731a5f2
[3] ihttps://review.opendev.org/c/openstack/releases/+/772047

Change-Id: Ie2d055f171418e2d35db697dfb1aa17cd449bea9