Release final castellan for Wallaby
This proposed a release for castellan to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, March 8.

$ git log --oneline --no-merges 3.7.0..35f1a20
35f1a20 Fix assertion typo in barbican key manager unit test
f638cdf Dropping lower constraints testing
4a4544b Don't expect barbican service name to be barbican.
47b1733 Use py3 as the default runtime for tox

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Id73da4b9db2ed57c06ebf9d541f8baee5b5003d4