Release final kuryr for Wallaby
This proposed a release for kuryr to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, March 8.

$ git log --oneline --no-merges 2.2.0..f973749
cdb9fe3 Remove run_as_root which fails without root_helper
78c31d3 Remove the unused coding style modules

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ied58267663edeecfc970adbb4624c89a68937896