Release final neutron-lib for Wallaby
This proposed a release for neutron-lib to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, March 8.

$ git log --oneline --no-merges 2.9.0..698e4c8
698e4c8 Add PLACEMENT_REPORT plugin constant

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I2d354656640a34ba31537ac0790d8989fa402aef