Release final openstacksdk for Wallaby
This proposed a release for openstacksdk to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, March 8.

$ git log --oneline --no-merges 0.53.0..164231a0
e30fd860 Improve Ironic API on OpenStack SDK
c81c1f61 Fix a trivial error in one of the error messages
5f952298 Add shared file systems support
3371e2b0 Add compute microversion 2.70
e9dde292 Add TODO to remove md5 wrapper
7040bf8c Remove unnecessary string formatting
63c5f8a9 Drop .json suffix from networking URLs
10a1ee45 Support Deploy Templates for Ironic API
b25241cc Change microseconds to total_seconds()
8442aeab Move 'collections.Mapping' to 'collections.abc'
4a469fa5 Add ALPN support to load balancer pools
39d35297 Update Open Telekom Cloud vendor docu
30116cf3 Complete compute.hypervisor functions
27536928 encapsulate md5 calls for fips
78dd1a73 Support SNAP_REAL_HOME when using openstacksdk inside a snap package
b86edc36 Add id query parameter to sg rules

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: If535a6aaec14e1588af913b5705d39cfa46d6423