Release final os-vif for Wallaby
This proposed a release for os-vif to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, March 8.

$ git log --oneline --no-merges 2.3.0..6167658
44d8937 Resolve dependency issues
7b54dec Drop use of deprecated collections classes

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I9a1c3d06521fc7c562caf63bae664c22b2d9306c