Release final oslo.config for Wallaby
This proposed a release for oslo.config to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, March 8.

$ git log --oneline --no-merges 8.4.0..cfa2564
cfa2564 Bump minimum version for PyYAML to 5.1
50723fa remove lower-constraints.txt
05b6a33 inherit from object is not required for py3
401dcd9 Move zuul config file to root folder
cefa4d4 Dropping lower constraints testing
f1f8730 Remove Babel from lower-constraints.txt

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Idbec8dc7663db40f3cda029e5b588071edc83f3b