Release final oslo.context for Wallaby
This proposed a release for oslo.context to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, March 8.

$ git log --oneline --no-merges 3.1.1..0d02866
0d02866 Switch to collections.abc.MutableMapping
3542bea Dropping lower constraints testing
b6a92c9 Add Python3 wallaby unit tests
747e723 Update master for stable/victoria
3b6da99 Adding pre-commit

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I4bca5c0ae21428d83dcd156d05bcd49b22f2f631