Release final ovsdbapp for Wallaby
This proposed a release for ovsdbapp to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, March 8.

$ git log --oneline --no-merges 1.8.0..d6383a5
13b4dd1 Log the name of the command in execute()
26f24cf Drop py37 and add py38 test

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I2029bb7d2cfee207b4f71d881ce249066570560b