Release final sushy for Wallaby
This proposed a release for sushy to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, March 8.

$ git log --oneline --no-merges 3.6.1..d4c7e03
375b29b Follow up TaskMonitor refactor
7a9a2b6 Refactor TaskMonitor and update Volume methods
07ca7f4 Don't log ERROR if GET of /redfish/v1/SessionService fails
d9004ec Fix deprecation on collections.MutableMapping
09cee5c Automatically retry HTTP 5xx on GET requests
4abea18 Initial support for secure boot databases

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I1e6ee139d84ec49ea331028818995f63513fa599