Release final tosca-parser for Wallaby
This proposed a release for tosca-parser to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, March 8.

$ git log --oneline --no-merges 2.2.0..a1c2cbe
d353bbc Remove depreciated file babel.cfg
6f253c5 Dropping explicit unicode literal
1f6f7a0 Update TOX_CONSTRAINTS_FILE
ce8fa0f Move tosca-parser to new hacking 4.0.0
df16f01 Support "Implementation" definition as Artifact Name
b0c5c55 Drop Python 2 support
62f0d45 Add Python3 wallaby unit tests
00d3a39 Add validation for required property of policies
0653750 Fix error parsing scalar unit contraints

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ie1dd06525b161219de0d85ac8dad7aa0fcc8573c