Release final heat-translator for Wallaby
This proposed a release for heat-translator to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, March 8.

$ git log --oneline --no-merges 2.1.0..7109d4c
615c76b Remove depreciated file babel.cfg
e41b382 Dropping explicit unicode literal
91c4415 Update TOX_CONSTRAINTS_FILE
6687b35 Move heat-translator to new hacking 4.0.0
4809045 Remove six support
1823e85 Switch to wallaby job template
828ede5 [goal] Migrate testing to ubuntu focal
bfb642e drop mock from lower-constraints
7dc0b74 Fix hacking min version to 3.0.1

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I3d16919d8ee19c2160fcccd5a7bef35d50fc488a