Release final python-ironicclient for Wallaby
This proposed a release for python-ironicclient to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, March 11.

$ git log --oneline --no-merges 4.6.0..59ffddd
f8af1df Add tests for 'baremetal port create' command

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Id3967320bff1f37f4705228359d8bf85b26db080