Release final python-magnumclient for Wallaby
This proposed a release for python-magnumclient to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, March 11.

$ git log --oneline --no-merges 3.3.0..e2bcf6e
e2bcf6e Uncap PrettyTable
201b352 Add CT tags argument and field to client
0799acf Nodegroup min_node_count should default to 0
0a35b68 Drop lower constraints testing

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ibff157e8eeb04c50aaeb1a04b9c45c8fa23040be