Release final python-monascaclient for Wallaby
This proposed a release for python-monascaclient to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, March 11.

$ git log --oneline --no-merges 2.2.1..75e44a1
75e44a1 Uncap PrettyTable
952c4c1 Fix zuul publish docker image job
45df91a Add doc/requirements
43203ea Update doc8 version
feb2214 Remove six
20e1d9f Add Python3 wallaby unit tests
7f12494 Update master for stable/victoria

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ib1e04d375b1f42b1ab74edca6ab350c1ca174be5