Release final python-octaviaclient for Wallaby
This proposed a release for python-octaviaclient to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Thursday, March 11.

$ git log --oneline --no-merges 2.2.0..6dffa41
dfcdd57 Improve error messages in list/show calls
80a5331 Remove install unnecessary packages
21f1d67 Add ALPN support for pools
381c2fc Fix lower-constraints job
ca9146b remove unicode from code
9c4b85a Add tag support for Octavia resources
f832587 Add support for PROXYV2 protocol
1a133a3 Add Python3 wallaby unit tests
1d34fcc Update master for stable/victoria

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I145846fb5c4dbd33c3f28700bee9d8f05f7296f2