Release final python-senlinclient for Wallaby
This proposed a release for python-senlinclient to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, March 11.

$ git log --oneline --no-merges 2.2.0..e6edaee
e6edaee Remove unicode from python client

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Icaadc146f5608e0de7bdab672a3fd4fa4858b028