Release final python-swiftclient for Wallaby
This proposed a release for python-swiftclient to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, March 11.

$ git log --oneline --no-merges 3.11.0..06b36ae
06b36ae Drop lower-constraints testing
e687636 remove unicode from code

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I0eb722fe343f191a60630a68bb7048a90dbf3fef