Release final python-tackerclient for Wallaby
This proposed a release for python-tackerclient to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, March 11.

$ git log --oneline --no-merges 1.5.0..4e6dc4c
4e6dc4c CLI for Individual VNF LCM Operation Occurrence
6b29bb7 Support CLI for Getting List of VNF LCM Operation Occurrences
1299e15 Support of Retry VNF command in openstackclient
cfe1411 Support CLI of Change External VNF Connectivity
5816753 Support of Fail VNF command in openstackclient
e7a0c96 Fix error message for nonexistent vnf package
68df829 Dropping explicit unicode literal
17f4bca Fix old links in installation guide
69d9668 Update TOX_CONSTRAINTS_FILE

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I84ada4201cd05b2d65c8c28d5cb7208f7700e038