Release final python-vitrageclient for Wallaby
This proposed a release for python-vitrageclient to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, March 11.

$ git log --oneline --no-merges 4.2.0..8293be5
8293be5 Revert "remove py37"
764e430 remove py37
7dbc1e8 Remove unicode  from client

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ia6be75df1986537815f2438a7fac4ee4e522d799