Release final python-novaclient for Wallaby
This proposed a release for python-novaclient to get any final merged changes
out prior to the lib freeze for Wallaby going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Monday, March 11.

$ git log --oneline --no-merges 17.3.0..54d4da11
54d4da11 Add support for microversion v2.88
e4595392 requirements: Remove simplejson
f50782f4 Uncap PrettyTable

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I47874429617678a398d3fd40aa9113dd3dbf5dc2