Propose to move pycadf to independent
Project did not have any change merged over the cycle, maybe it is time
to transition it independent.

If the team is ready to proceed, please leave a +1 to acknowledge this
transition and we will process right away. If the team doesn't agree,
leave a -1 with some detail on the status, then we will cut the Wallaby
stable branch for this project.

Change-Id: I34c37265237ec5ddb544ff7454f11cc45a1d87d1