[mistral] Create for all client and non-client lib wallaby branch
This creates the stable/wallaby branch for all client and non-client libraries
now that we are past the lib freeze on March 11.

If a team wants to wait for a specific patch to make it to the library,
someone from the team can -1 the patch to have it held,or update that
patch with a different commit SHA.

Patches with no response will be merged by Friday, March 19.

Change-Id: Ibff60e6a94654e76befd9fa4fdaebe39389c0903