[cyborg] Create for all client and non-client lib wallaby branch
This creates the stable/wallaby branch for all client and non-client
libraries
now that we are past the lib freeze on March 11.

If a team wants to wait for a specific patch to make it to the library,
someone from the team can -1 the patch to have it held,or update that
patch with a different commit SHA.

Patches with no response will be merged by Friday, March 19.

Change-Id: I58a9c1e8e2117931f44f56a377b4e2267a6ee459