Abandon monasca-transform
This project have been tagged as deprecated 6 months ago and a patch to
abandon it have been proposed on the governance repo [1].

However, looks like deprecation was not properly marked in governance.
It means was is not done completely that is why release things did not get.

gmann proposed to follow up with Monasca team if that is retired or
deprecated and update governance reference too.

We can remove it from release and the TC will take care of
governance things properly after discussion with Monasca team.

According to our discussion with the TC these changes simply remove the
deliverable for Wallaby.

The project will be no longer released.

[1] https://review.opendev.org/c/openstack/governance/+/783624

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I6b9f419080afd534f25fcdf484dd4f22b90619a5