Proposing ironic-prometheus-exporter for wallaby
Proposing release for cycle-with-intermediary deliverables that have not released yet.

If you wants to wait for a specific patch to make it to the release, someone from your team can -1 the patch to have it held, or update that patch with a different commit SHA.

Patches with no response will be merged by Friday March 26.

$ git log --oneline --no-merges 2.1.1..1ec7859
1ec7859 Add bifrost job and update docs
198f07d Reconfigure tox.ini
36f2ccb Remove lower-constraints job
c5d4ea9 Set safe version of hacking
e33f2e7 Add Python3 wallaby unit tests
0f9da06 Migrate IPE job to focal
8d25757 Update master for stable/victoria

Signed-off-by: Armstrong Foundjem <foundjem@ieee.org>
Change-Id: Ib12dbbb007c93497776623a4c4b6d79249376a2a