Release kuryr-tempest-plugin for wallaby
This is the automatic release for kuryr-tempest-plugin.
Please +1 if this is ready to go, or -1 if the team needs a little more
time. If so, please update with the new commit hash to use when ready.

$ git log --oneline --no-merges 0.8.0..36fd12e
36fd12e Fix create_vm_for_connectivity_test
34fe966 Adds connectivity test for SCTP service
8f24190 Increase the timeout for controller to be ready
0412a57 Make stable/train job non-voting
e38e432 Updates Kuryr tempest README
60f7788 Adds Kuryr-sctp-demo application
a646091 Make Stein gate non-voting
1eaffa9 New test for services without selectors with NP.
e564b66 Change registry from docker.io to quay.io for test container image.
202409c Remove uselsess files and tests

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I7dcf9d0b8e6f7a8109cb494d027124ec300dc746