Release senlin-tempest-plugin for wallaby
This is the automatic release for senlin-tempest-plugin.
Please +1 if this is ready to go, or -1 if the team needs a little more
time. If so, please update with the new commit hash to use when ready.

$ git log --oneline --no-merges 1.2.0..652bff9
9724ee7 negative cases for update receiver
57ec4df updates a receiver
cc30e0c negative cases for update action
44c54cb Rmove six forgotten before
81c87af list profile type operation
a9b828c Add stable branch jobs on the plugins master gate

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I2e37cffcf2a795372a9fa21c82dc01b06ab7b4f0