Clarifying how to deal with tempest-plugin during series
Previously the tempest plugins were driven by the legacy
`cycle-automatic` model. However this model have been abandoned.

The cycle-automatic release model is now better described by the
`cycle-with-intermediary` model combined with `stable-branch-type: none`

The `cycle-automatic` model was used by specific technical deliverables
that needed to be automatically released once at the end of a cycle.

the tempest-plugin are branchless and `cycle-with-intermediary`
however our process doesn't handle them at the end of the cycle
like did the `cycle-automatic`.

These changes clarifying that point.

Change-Id: I6b3619a44d0aea1fa4852e650cbafcb08c853353