Moving OSA roles to trailing projects
OpenstackAnsible is a trailing project so we think that these roles
should be aligned with the parent project of OSA team.

During previous series these roles and the parent project have
been branched more or less at the same time [1][2].

These changes propose to embrace the trailing nature of these
roles to avoid confusing at the end of each series.

[1] https://opendev.org/openstack/releases/commit/d35b5cce08ff8ffe44a20e4928013e6466dda761
[2] https://opendev.org/openstack/releases/commit/e838bcf0036c5814aa0d9962d8fd1d49967af619

Change-Id: I43381fa4becc006b64482ba0e16cd1b19e556e62