Final RC patch for cyborg
This creates a new RC before the final release candidate deadline to
release the latest fixes and/or translations.

Please ack with a +1, or if there is some reason to hold off, leave a -1.

$ git log --oneline --no-merges 6.0.0.0rc1..7fbae95
7fbae95 Remove personal name in the docs
4040843 Revert device_profile key to show device profile API
8ac560b tools: Allow check-cherry-picks.sh to be disabled by an env var
cbb9857 [stable-only] Update TOX_CONSTRAINTS_FILE for stable/wallaby
cc87094 [stable-only] Update .gitreview for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I8c594eb269c015101aceeb2ed7e02ad5343b0329
Depends-On: https://review.opendev.org/c/openstack/cyborg/+/785395