Allow to change the used branch
The branch to use in process_auto_release is determined by the
branch linked to a series's status, however, at some point during
a series projects are branched and the status isn't yet updated
so used branch refer to master, and so, to the next series.

During the final-rc delta of changes between RC1 and RC2 is
retrieved from master, which is wrong.

These changes allow to choose another branch if needed.

The default used branch will remain the series retrieved
by status.

Change-Id: Iff8df80aa6f3ba67c04e0674d7c72474f3a4e73a