Wallaby final releases for cycle-with-rc projects
This tags the final release for all cycle-with-rc projects for Wallaby.
While not required, having PTLs ack this patch would be nice to include
that record in the merged metadata.

Release activity should now be frozen until this is processed on Apr 14.
Any bugfixes merged to stable/wallaby can be queued up and ready to go
as a follow on stable release after the coordinated release date.

There are different diff-start values for each project. This is used
when generating release notes to denote the point where the
stable/victoria branch started to diverge from what is included on
stable/wallaby. For some, this will be the final Victoria major version,
but if multiple RCs were done, this may show the initial RC1 release
version. This is normal.

Change-Id: Ia6d79df224a792915976864857fe2271e031b54d