Recreate puppet-freezer release with the consistent version
The previous release patch used a wrong tag, which resulted in
inconsistency between version written in metadata and one in tag.
This change recreates the release using the consistent version.
Since we have already created 18.4.0 tag accidantally, this change
bumps the major version to 18 to avoid decreasing major version in
any future release.

Change-Id: I8d775f3aea693ea938f40340fe16b6c6fd864f6e