Move tooz to the independent model
During Wallaby the oslo team decided to move tooz to the independent
release model, however, some releases have been done in the mid of the
cycle so we prefered to keep it the current coordinated series (Wallaby
at this time).

Tooz is mature and stable enough, all the maintainers agreed with this
proposition, so, It doesn't make sense to retains tooz as a series based
deliverable anymore.

Change-Id: Idb6b4994bdd9bb566d7ddc8eaf9d145738f18447