ironic-python-agent-builder release model change
During the recent xena ptg, the ironic community had a discussion about
the need to move the ironic-python-agent-builder project from an
independent model to the standard release model.

When we initially split the builder from ironic-python-agent, we decided
against it, but considering some problems we encountered during the
road, the ironic community seems to be in favor of the change.
The reasons for this are mainly to strictly align the image building
project to ironic-python-agent releases, and ease dealing with the
occasional upgrade of tinycore linux, the base image used to build the
"tinyipa" ironic-python-agent ramdisk.

We decided to split them at some point, and we'd like to keep them
separated, but we realized that it would make sense to keep them synced
by branch.

The ironic contributor suffered CI pain over the
past year where code handling distributions has broken several times.
Largely these breaks have had to be navigated around with specific
version pinning, and settings changes, because we were unable to fix
something a release or two back inside the tooling, in this case
ironic-python-agent-builder. In other words, if the ironic team had that
capability and those branches/tags, we would have been less in a world of
pain because we wouldn't have had to make changes on every job/branch
across every repository where we were hitting issues with ipa-b. Reduction
of pain is why this change is being proposed.

We should also notice that the Kolla team also use ipa-builder and they
faced similar issues with Kayobe.

These changes propose to cut ipa-builder for Wallaby in a retro active
manner (Wallaby is already out at this point).

These changes use the same model that the one used by
ironic-python-agent.

These changes create a new tag (2.7.0) based on the same hash that the
previous independent version (2.6.0), mostly to properly init the series
with a meaningful version number, avoid misleading with future bugfix
releases versionned with a minor bump, create the wallaby branch from the
right point, avoid errors with our tooling related to missing tags for a
series.

These changes also init ipa-builder for Xena.

http://lists.openstack.org/pipermail/openstack-discuss/2021-April/022032.html

Change-Id: Ia899b39dbc68b245e1b84518363de5420af500c0