Release cliff for Xena-1 milestone
This is a library release for oscliff for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 3.7.0..d562aae
2fdf358 setup.cfg: Replace dashes with underscores
0753451 Replace getargspec with getfullargspec
fa0535a Use py3 as the default runtime for tox
0062d33 Add Python3 xena unit tests

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I8d3c3c932dc5c91f5e0476e14030124307991ae2