Release heat-translator for Xena-1 milestone
This is a library release for osheat-translator for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 2.2.0..491348b
491348b Add Python3 xena unit tests
c6e0492 ETSI-NFV SOL 001 translation: AffinityRule
2c94431 Fix processing related to required properties

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I4d82603e13e2c6038276f60fcdedbd7d897c3b34