Release ironic-lib for Xena-1 milestone
This is a library release for osironic-lib for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 4.6.1..36b8486
36b8486 convert_image: add flags required for IPA
df9a2e8 Raise qemu-img memory limit to 2 GiB
37ac433 setup.cfg: Replace dashes with underscores
9fb5be3 Stop testing the iscsi deploy interface
4cf1b4f Remove runtime dependency on pbr
fb8b6c1 Add Python3 xena unit tests

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ic9099ec98cb08374a6abcdd06f0374eb126716a6