Release os-brick for Xena-1 milestone
This is a library release for osos-brick for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 4.3.0..f18a33b
f18a33b Introduce mypy
1b2e229 multipath/iscsi: remove devices from multipath monitoring
985a49b vmware: Use cookiejar from oslo.vmware client directly
534c04c setup.cfg: Replace dashes with underscores
eb566ce Enable bandit runs for tox
d4205bd iSCSI: Fix flushing after multipath cfg change
9afa1a0 Add Python3 xena unit tests
ab57392 Update master for stable/wallaby
7cfdb76 Dropping explicit unicode literal
b580124 connector: Fix comment of get_connector_properties regarding enforce_multipath
28545c7 RBD: catch read exceptions prior to modifying offset

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I82f3fcac8bd30c48dce2effcb3f8a59be919de6b