Release os-ken for Xena-1 milestone
This is a library release for osos-ken for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 1.4.0..5312d23b
5312d23b pdf-docs: use UPPER_CONSTRAINTS_FILE for doc build
0220f50e Add Python3 xena unit tests
9223eb9a Update master for stable/wallaby
bfffa92e Add requirements-check job for os-ken
cd4926b2 Remove os_ken/app and os_ken/services/protocols/bgp/api/jsonrpc modules

This will release a major version due to cd4926b2 which remove os_ken/app and
os_ken/services/protocols/bgp/api/jsonrpc modules

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: If29515c42b00b34bac9ee587cf97d0b1cd83f021