Release os-vif for Xena-1 milestone
This is a library release for osos-vif for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 2.4.0..e937367
c9b2077 Creating oslo.config.opts entry_points for plugins
fcfa244 setup.cfg: Replace dashes with underscores
d8af356 Add Python3 xena unit tests
e1ad9ea Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I28761067b2dba607cff21ade09e4bb58a09d10eb