Release osc-placement for Xena-1 milestone
This is a library release for ososc-placement for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 2.2.0..37e47af
39befda Verify result for inventory set --dry-run
a4ac717 Add "--resource-class" to  allocation unset
9778535 Note env OS_PLACEMENT_API_VERSION support
05ca1ee Switch default to use latest microversion
8ac8c86 Support auto-negotiated microversion
6914082 setup.cfg: Replace dashes with underscores
ee98b03 Update master for stable/wallaby
714d4e6 Add openstackclient-plugin-jobs
8def9e8 Add functional-py39 tox target
e27453e Use TOX_CONSTRAINTS_FILE
c16bff5 Add py38 package metadata
826bc1f Add Python3 wallaby unit tests
93eb4e8 Update master for stable/victoria
eb5a625 Support granular allocation candidate list
db98792 Support multiple member_of query parameter
dee5bff Include usage in 'inventory list', 'inventory show'

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ib3650b06afb0e59d8473c0475c8fac5307ffb42b