Release oslo.cache for Xena-1 milestone
This is a library release for ososlo.cache for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 2.7.0..d89cc82
d89cc82 setup.cfg: Replace dashes with underscores
b1528f9 Move flake8 as a pre-commit local target.
68917b1 Add Python3 xena unit tests
83a443f Update master for stable/wallaby
6f3c6c9 Adding technical details about the memcache pool backend.
e5d0b20 Add bug comment and fix nits
9606e85 Dropping lower constraints testing.
51d7b46 Drop use of deprecated collections classes
0cf7b0e Use py3 as the default runtime for tox

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I709018dc678544915458aa8cbad8582956d32958