Release oslo.context for Xena-1 milestone
This is a library release for ososlo.context for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 3.2.0..64842a1
64842a1 setup.cfg: Replace dashes with underscores
655daf1 Fix formatting of release list
ceca2bb Move flake8 as a pre-commit local target.
16c7382 Add Python3 xena unit tests
dbdca84 Update master for stable/wallaby
a24a5a0 Remove lower-constraints remnants
67a4675 Drop use of deprecated collections classes
d4a4b6d Use TOX_CONSTRAINTS_FILE
f0da2bb Use py3 as the default runtime for tox

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I876c3e199360f39e63c39753545a60e3829a63e6