Release oslo.messaging for Xena-1 milestone
This is a library release for ososlo.messaging for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 12.7.1..355e0bd7
355e0bd7 Remove references to 'sys.version_info'
a5ad998b Fix formatting of release list
f43a7068 Move flake8 as a pre-commit local target.
83d0e453 Add Python3 xena unit tests
bddf5310 Update master for stable/wallaby
e52dc2ac Remove lower constraints.
8d29ab3d bindep: Add 'librdkafka-dev' dependency

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I6389de7eb3be8778839ecb300b07d5b80d9551cb