Release oslo.upgradecheck for Xena-1 milestone
This is a library release for ososlo.upgradecheck for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 1.3.0..3ca8698
75fe376 Move flake8 as a pre-commit local target.
276ecd7 Add Python3 xena unit tests
db3e3a9 Update master for stable/wallaby
8a43c37 Remove lower-constraints remnants
f234316 Uncap PrettyTable
456bd85 Remove six
a962943 Dropping lower constraints testing
20b199e Use TOX_CONSTRAINTS_FILE
09c42be Use py3 as the default runtime for tox

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I08b56d3dac10382af01b02dc8240ac4ad9734e6b