Release python-designateclient for Xena-1 milestone
This is a library release for ospython-designateclient for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 4.2.0..9bc517a
820f1e9 Remove six and update lower-constraints appdirs
f8353c5 Add Python3 xena unit tests
cafa8a3 Update master for stable/wallaby
0e32928 Add Python3 wallaby unit tests
1f6f489 Remove install unnecessary packages
e3b88ed Fix lower-constrains job
8492a9e Update readme.rst
c0f62f2 bump py37 to py38 in tox.ini
196cb61 Fixed zone transfer accept list command
dd7fd96 Update master for stable/victoria

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I845d699e26cfff3d63e8c588a64202c92736824f