Release python-ironicclient for Xena-1 milestone
This is a library release for ospython-ironicclient for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 4.6.1..384bd6e
384bd6e setup.cfg: Replace dashes with underscores
4a22f29 Make baremetal --debug actually work
ac718d5 Fix the functional tests
70143bc Update master for stable/wallaby
be2f116 Fix --fields network_data

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I4179c48eb5ce19676c3e511410348d853d7705ee