Release python-magnumclient for Xena-1 milestone
This is a library release for ospython-magnumclient for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 3.4.0..ab601bd
2827533 Fix master_lb_enabled for cluster creation
f7551a6 Fix failing to parse json error msg

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I7a1a49e8b4ebd16775780d2558abd6935476f9f7