Release python-novaclient for Xena-1 milestone
This is a library release for ospython-novaclient for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 17.4.0..f431295d
665f1c8d Change minversion of tox to 3.18.0
935fe75a Refactor constructing request body
3dc9ad97 setup.cfg: Replace dashes with underscores
b5faf37f Use py3 as the default runtime for tox
21edd11b Add unit tests for client logger
d4399d3d When creating a client, pass the default logger
ef6b853f Add Python3 xena unit tests
91df1949 Update master for stable/wallaby
8cdbb0f9 Use well named anchor into the microversion history

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I8f4aea2fcfbb3a7c52ae3c30a19ee18ca5f94234