Release python-swiftclient for Xena-1 milestone
This is a library release for ospython-swiftclient for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 3.11.1..38c435f
2beddc2 setup.cfg: Replace dashes with underscores
58f0700 Fix Typo in shell.py
998bb8b Use py3 as the default runtime for tox
6966fbe Add Python3 xena unit tests
c8b48de Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: Ifb48e6d4d147a7c4b4a1835a47f20355bcb45355