Release python-tackerclient for Xena-1 milestone
This is a library release for ospython-tackerclient for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 1.6.0..b5603cc
b5603cc Use assertCountEqual instead of assertItemsEqual
a7ee00e Add Python3 xena unit tests
785c6db Adds support force delete VNFFG
124be38 Fix failing UT in TestListVnfLcmOp

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: If571266fe5839425d33dc3d5d7824f8bfd81fed3