Release python-troveclient for Xena-1 milestone
This is a library release for ospython-troveclient for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 7.0.0..cd68986
cd68986 Get backup list by instance ID or name
e41d08d Support project name in quota CLI
1d53377 Respect endpoint, cacert and debug options
8a3a26e Add Python3 xena unit tests
23d3ed7 Update master for stable/wallaby

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: If593d0e63af2de32e11ecbb80ea6a0de3fc96541