Release tosca-parser for Xena-1 milestone
This is a library release for ostosca-parser for the Xena-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 2.3.0..88988bb
424cc36 Simplify log translations
5860c6f Update validation of function `get_operation_output`
093fc80 Make custom definitions biggest priority

Signed-off-by: Hervé Beraud <hberaud@redhat.com>
Change-Id: I5a5aed8d4ff103af59e5e43e12d825416aff8e2c